requirements.txt
python 3.10.6
openpyxl-3.0.10
numpy-1.23.1
pandas-1.4.3 python-dateutil-2.8.2 pytz-2022.1 six-1.16.0